+++ src/dependency_injector/__init__.py
@@ -1,6 +1,6 @@
 """Dependency injector top-level package."""
 
-__version__ = '3.38.1'
+__version__ = '3.39.0'
 """Version number that follows semantic versioning.
 
 :type: str