+++ docs/main/changelog.rst
@@ -7,6 +7,14 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+3.39.0
+------
+- Add application examples with single and multiple containers.
+- Remove "Services" application examples.
+- Split examples page into "Examples" with main examples and "Other Examples" with secondary
+  examples.
+- Move "Installation" page to "Introduction" section.
+
 3.38.1
 ------
 - Fix README.